+++ EmeraldIOS/Components/TextfieldView/Constant/ConstantIdentifier.swift
@@ -1,0 +1,14 @@
+//
+//  ConstantIdentifier.swift
+//  EmeraldIOS
+//
+//  Created by Carlos Pava on 10/12/19.
+//  Copyright © 2019 Condor Labs. All rights reserved.
+//
+
+import Foundation
+
+struct ConstantIdentifier {
+    static let textField = "emeraldTextField"
+    static let errorLabel = "errorEmeraldLabel"
+}
